Flush Mount / Custom panel issue resolved
The code was modifying the Flush panel parameter. Since this parameter was controlled by the Surface mount parameter, the flush panel parameter was a Read-Only parameter. Therefore the code was failing.

Solution: A try catch statement was added to the SetInt method of the Flush Mount parameter set method, whereby, if it tried to set the Flush mount method and it fails, the code will try to set the value (opposite value*) for Surface mount.

*Note: Opposite value, because Flush mount = not(Surface mount) => Formula in the family